LayoutTests/userscripts/script-not-run-for-fragments-expected.txt

CONSOLE MESSAGE: line 0: injected
CONSOLE MESSAGE: line 0: injected
This test should not infinitely inject user scripts. 
